Isolate MerciText component
